infra/requirements.txt
# CDK dependencies
aws-cdk-lib>=2.80.0
constructs>=10.2.69
python-dotenv>=1.0.0

# LocalStack dependencies
awscli-local>=0.20
localstack>=2.1.0

# Build dependencies
pyinstaller>=5.13.0
typer>=0.9.0
rich>=13.3.5
tabulate>=0.9.0
